Move query building to new class
